Back out some recent changes
Removed the cloudfoundry integration dependency (so we are not
going to be a service broker) because it added security to the
/v2/* endpoints and clients didn't seem to like that.

Removed the "ignore 404s" flag from the Jersey filter. It was there
to allow Spring MVC to handle the missing resources in /v2/*, but
seems to result in a lot of 405s when clients try to push changes.

See gh-26, gh-27